requirements.txt
# Configuration file support
PyYAML>=6.0
jsonschema>=4.0

# Data analysis
pandas>=1.3.0
numpy>=1.20.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Logging and utilities
python-dateutil>=2.8.0

# === Local LLM Support ===

# Ollama Python SDK
ollama>=0.1.0

# LangChain for provider abstraction
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0

# === Fine-Tuning Support ===

# Core ML frameworks
transformers>=4.35.0
torch>=2.0.0

# PEFT/LoRA for efficient fine-tuning
peft>=0.7.0

# Training acceleration
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Supervised fine-tuning
trl>=0.7.0

# Dataset management
datasets>=2.14.0

# === Code Parsing for Dataset Generation ===

# Tree-sitter for code parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0

# === Utilities ===

# Rich terminal output
rich>=13.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.32.0

# Data validation
pydantic>=2.0.0

# API clients
anthropic>=0.31.0

# Note: For GPU support, install CUDA-compatible PyTorch:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
